# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Scan -\u003e Plan -\u003e Approve -\u003e Fix -\u003e Rollback REST API for golangci-lint issue triage.",
        "title": "Golangci Lint Dashboard API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/rollbacks": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Rule BR-4: revert-only, never reset/force-push. Fails if the fix made no commit, or was already rolled back.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rollbacks"
                ],
                "summary": "Roll back an applied fix",
                "parameters": [
                    {
                        "description": "Fix to roll back",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.createRollbackRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "fix made no commit -- nothing to roll back",
                        "schema": {
                            "$ref": "#/definitions/golangci_backend_frontiir_utils.RestErr"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/golangci_backend_frontiir_utils.RestErr"
                        }
                    },
                    "409": {
                        "description": "already rolled back, or a scan/fix already running for this repo+branch",
                        "schema": {
                            "$ref": "#/definitions/golangci_backend_frontiir_utils.RestErr"
                        }
                    }
                }
            }
        },
        "/rollbacks/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rollbacks"
                ],
                "summary": "Get a rollback's result by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Rollback ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/storage.RollbackHistory"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/golangci_backend_frontiir_utils.RestErr"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.createRollbackRequest": {
            "type": "object",
            "required": [
                "fix_history_id"
            ],
            "properties": {
                "fix_history_id": {
                    "type": "string"
                }
            }
        },
...
        "golangci_backend_frontiir_utils.RestErr": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Machine-readable business error code (e.g. \"SLA_EXPIRED\")",
                    "type": "string"
                },
                "data": {
                    "description": "Optional structured payload (e.g. conflicting ticket detail)"
                },
                "message": {
                    "description": "User-friendly error message",
                    "type": "string"
                },
                "status": {
                    "description": "HTTP status code",
                    "type": "integer"
                }
            }
        },
...
        "storage.RollbackHistory": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "fix_history_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "post_rollback_scan_id": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "result": {
                    "description": "reverting|done|conflict|failed -- added when POST /rollbacks was\nimplemented: 11-sequence-diagram.md's own Flow 4 describes two\noutcomes (revert succeeds vs conflicts) that need to be\ndistinguishable, and apply is worker-queued like scan/plan/fix so\nneeds an in-flight state too, but no such field existed.",
                    "type": "string"
                },
                "revert_commit_sha": {
                    "type": "string"
                },
                "rolled_back_by": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}